cleanup all code
