+++ CMakeLists.txt
@@ -4,8 +4,12 @@ project(main CXX)
 set(CMAKE_CXX_STANDARD 17)
 set(SW_BUILD_SHARED_LIBS 1)
 set(DEPENDENCIES
+    org.sw.demo.boost.serialization-1.75.0
+    org.sw.demo.boost.core-1.75.0
     org.sw.demo.wxwidgets.core-3.1.4
     org.sw.demo.wxwidgets.base-3.1.4
+    org.sw.demo.google.tesseract.libtesseract-master
+    org.sw.demo.danbloomberg.leptonica-1.80.0
 )
 
 find_package(SW REQUIRED)
@@ -13,8 +17,9 @@ sw_add_package(${DEPENDENCIES})
 sw_execute()
 
 # add include dir
-include_directories(./)
+include_directories(include/)
+file(GLOB SOURCES "./src/*.cpp")
 
-add_executable(main WIN32 mainApp.cpp)
+add_executable(main WIN32 ${SOURCES})
 target_link_libraries(main ${DEPENDENCIES})
