+++ include/mainApp.h
@@ -1,5 +1,5 @@
 #pragma once
-
+#include <wxcmake.h>
 #include <wx/wxprec.h>
 
 class mainApp : public wxApp
